src/test/java/kitchenpos/menus/tobe/domain/MenuGroupTest.java
package kitchenpos.menus.tobe.domain;

import static kitchenpos.global.TobeFixtures.menuGroup;
import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertAll;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

class MenuGroupTest {

    @DisplayName("메뉴 그룹을 등록할 수 있다.")
    @Test
    void create() {
        MenuGroup expected = menuGroup("두마리메뉴");
        MenuGroup actual = menuGroup("두마리메뉴");

        assertThat(actual).isNotNull();
        assertAll(
                () -> assertThat(actual.getId()).isNotNull(),
                () -> assertThat(actual.getName()).isEqualTo(expected.getName())
        );
    }
}
